requirements.txt
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
deep-translator==1.11.4
idna==3.6
packaging==23.2
Pillow==10.1.0
pytesseract==0.3.10
requests==2.31.0
soupsieve==2.5
urllib3==2.1.0
